Fix default CMake build
